move containing scene helper
